+++ src/api/math.ts
@@ -7,7 +7,6 @@ import {
     ExtendedAssetsConfig,
     ExtendedAssetsData,
     MasterConstants,
-    PoolAssetConfig,
     PoolConfig
 } from '../types/Master';
 import { Dictionary } from '@ton/core';
@@ -20,6 +19,7 @@ import {
     UserBalance
 } from '../types/User';
 import { UNDEFINED_ASSET } from '../constants/assets';
+import { addReserve, isBadDebt } from './liquidation';
 
 export function mulFactor(decimal: bigint, a: bigint, b: bigint): bigint {
     return (a * b) / decimal;
@@ -294,19 +294,6 @@ export function presentValue(sRate: bigint, bRate: bigint, principalValue: bigin
 }
 
 /**
- * Determines if the provided pair of assets forms a bad debt.
- * @param totalSupply total supply worth amount
- * @param totalBorrow total borrow worth amount
- * @param liquidationBonus collateral liquidation bonus value
- * @param masterConstants pool constants
- */
-export function isBadDebt(totalSupply: bigint, totalBorrow: bigint,
-                          liquidationBonus: bigint, masterConstants: MasterConstants
-) {
-    return totalSupply * masterConstants.ASSET_LIQUIDATION_BONUS_SCALE < totalBorrow * liquidationBonus;
-}
-
-/**
  * Calculates health parameters of the specified user account based on its parameters
  * @param parameters
  */
@@ -361,232 +348,6 @@ export function calculateHealthParams(parameters: HealthParamsArgs) {
         isLiquidatable: _isLiquidable,
         isBadDebt: _isBadDebt
     };
-}
-
-/**
- * Check if the user is subject to liquidation
- * @param args
- */
-export function isLiquidatable(args: {
-    assetsData: ExtendedAssetsData, assetsConfig: ExtendedAssetsConfig,
-    principals: Dictionary<bigint, bigint>, prices: Dictionary<bigint, bigint>,
-    poolConfig: PoolConfig
-}): boolean {
-    const { isLiquidatable: res } = calculateHealthParams(args);
-    return res;
-}
-
-/**
- * Adds reserve to liquidation amount
- * @param amount raw liquidation amount
- * @param reserveFactor asset reserve factor
- * @param factorScale asset reserve factor scale
- * @returns liquidation amount with reserve
- */
-export function addReserve(amount: bigint, reserveFactor: bigint, factorScale: bigint): bigint {
-    return amount * factorScale / (factorScale - reserveFactor);
-}
-
-/**
- * Deducts reserve from liquidation amount
- * @param amount liquidation amount with reserve
- * @param reserveFactor asset reserve factor
- * @param reserveFactorScale asset reserve factor scale
- * @returns liquidation amount without reserve
- */
-export function deductReserve(amount: bigint, reserveFactor: bigint, reserveFactorScale: bigint): bigint {
-    return amount * (reserveFactorScale - reserveFactor) / reserveFactorScale;
-}
-
-/**
- * Converts worth value to specified asset amount
- * @param worth worth value (decimals = 9)
- * @param scale asset scale (10^asset_decimals)
- * @param price asset price
- */
-export function toAssetAmount(worth: bigint, scale: bigint, price: bigint): bigint {
-    return worth * scale / price;
-}
-
-/**
- * Converts calculates worth value of specified asset amount
- * @param amount worth value (decimals = 9)
- * @param scale asset scale (10^asset_decimals)
- * @param price asset price
- */
-export function toAssetWorth(amount: bigint, scale: bigint, price: bigint): bigint {
-    return amount * price / scale;
-}
-
-/**
- * Adds liquidation bonus to reward value.
- * @param value reward asset value
- * @param bonus liquidation bonus factor
- * @param scale liquidation bonus scale
- */
-export function addLiquidationBonus(value: bigint, bonus: bigint, scale: bigint): bigint {
-    return value * bonus / scale;
-}
-
-/**
- * Deducts liquidation bonus from reward value.
- * @param value reward asset value
- * @param bonus liquidation bonus factor
- * @param scale liquidation bonus scale
- */
-export function deductLiquidationBonus(value: bigint, bonus: bigint, scale: bigint): bigint {
-    return value * scale / bonus;
-}
-
-export type SuccessType = { ok: true };
-export type FailType = { ok: false };
-
-export type PreparedAssetInfo = {
-    config: AssetConfig,
-    data: AssetData,
-    price: bigint,
-    principal: bigint,
-    scale: bigint,
-    liquidationReserveFactor: bigint,
-    liquidationBonus: bigint,
-    present: UserBalance,
-    balance: bigint,
-    dust: bigint,
-};
-
-export type PreparedAssetInfoResult = (PreparedAssetInfo & SuccessType) | FailType;
-
-/**
- * Prepares extended asset info
- * @param assetId
- * @param assetsConfigDict
- * @param assetsDataDict
- * @param pricesDict
- * @param principalsDict
- * @param masterConstants
- */
-export function prepareAssetInfo(assetId: bigint,
-                                 assetsConfigDict: ExtendedAssetsConfig,
-                                 assetsDataDict: ExtendedAssetsData,
-                                 pricesDict: Dictionary<bigint, bigint>,
-                                 principalsDict: Dictionary<bigint, bigint>,
-                                 masterConstants: MasterConstants
-): PreparedAssetInfoResult {
-    if (!assetsConfigDict.has(assetId) ||
-        !assetsDataDict.has(assetId) ||
-        !pricesDict.has(assetId) ||
-        !principalsDict.has(assetId)) {
-        return { ok: false };
-    }
-
-    const config = assetsConfigDict.get(assetId)!;
-    const data = assetsDataDict.get(assetId)!;
-    const base = {
-        ok: true,
-        config, data,
-        price: pricesDict.get(assetId)!,
-        principal: principalsDict.get(assetId)!,
-        scale: 10n ** config.decimals,
-        liquidationReserveFactor: config.liquidationReserveFactor,
-        liquidationBonus: config.liquidationBonus
-    };
-
-    const assetPresent = presentValue(data.sRate, data.bRate, base.principal, masterConstants);
-    const dustPresent = presentValue(data.sRate, data.bRate, config.dust, masterConstants);
-
-    return { ...base, present: assetPresent, balance: assetPresent.amount, dust: dustPresent.amount };
-}
-
-/**
- * Calculates liquidation amount and corresponding collateral amount
- * @param supplyAmount user total supply worth amount
- * @param borrowAmount user total borrow worth amount
- * @param masterConstants evaa master contract constants
- * @param loanAsset loan asset pool config
- * @param collateralAsset collateral asset pool config
- * @param principalsDict user principals
- * @param assetsDataDict assets data collection
- * @param assetsConfigDict assets config collection
- * @param pricesDict assets prices
- * @returns maxLiquidationAmount max loan asset amount to transfer
- * @returns maxCollateralRewardAmount max collateral reward amount, which can be obtained
- */
-export function calculateLiquidationAmounts(
-    loanAsset: PoolAssetConfig,
-    collateralAsset: PoolAssetConfig,
-    supplyAmount: bigint,
-    borrowAmount: bigint, // from calculate health params
-    principalsDict: Dictionary<bigint, bigint>,
-    pricesDict: Dictionary<bigint, bigint>,
-    assetsDataDict: ExtendedAssetsData,
-    assetsConfigDict: ExtendedAssetsConfig,
-    masterConstants: MasterConstants
-): {
-    maxLiquidationAmount: bigint, maxCollateralRewardAmount: bigint
-} {
-    const loanInfo = prepareAssetInfo(loanAsset.assetId,
-        assetsConfigDict, assetsDataDict, pricesDict, principalsDict, masterConstants
-    );
-    const collateralInfo = prepareAssetInfo(collateralAsset.assetId,
-        assetsConfigDict, assetsDataDict, pricesDict, principalsDict, masterConstants
-    );
-
-    if (!loanInfo.ok || !collateralInfo.ok ||
-        loanInfo.present.type !== BalanceType.borrow ||
-        collateralInfo.present.type !== BalanceType.supply) {
-        return { maxLiquidationAmount: 0n, maxCollateralRewardAmount: 0n };
-    }
-
-    const liquidationBonusScale = masterConstants.ASSET_LIQUIDATION_BONUS_SCALE;
-    const collateralThreshold = masterConstants.COLLATERAL_WORTH_THRESHOLD;
-    const reserveFactorScale = masterConstants.ASSET_RESERVE_FACTOR_SCALE;
-    const reserveFactor = loanInfo.liquidationReserveFactor;
-    const liquidationBonus = collateralInfo.liquidationBonus;
-
-    let allowedCollateralAmount = collateralInfo.balance;
-    const _isBadDebt = isBadDebt(supplyAmount, borrowAmount, liquidationBonus, masterConstants);
-    if (!_isBadDebt) {
-        BigMath.max(collateralInfo.balance / 2n, BigMath.min(collateralInfo.balance, collateralThreshold));
-    }
-
-    const baseLiquidationWorth = BigMath.min(
-        toAssetWorth(loanInfo.balance, loanInfo.scale, loanInfo.price),
-        deductLiquidationBonus(
-            toAssetWorth(allowedCollateralAmount, collateralInfo.scale, collateralInfo.price),
-            liquidationBonus,
-            liquidationBonusScale
-        )
-    );
-
-    // calculate collateral amount
-    let collateralAmount = addLiquidationBonus(baseLiquidationWorth, liquidationBonus, liquidationBonusScale);
-    collateralAmount = toAssetAmount(collateralAmount, collateralInfo.scale, collateralInfo.price);
-
-    // calculate loan amount
-    let liquidationAmount = toAssetAmount(baseLiquidationWorth, loanInfo.scale, loanInfo.price);
-    liquidationAmount = addReserve(liquidationAmount, reserveFactor, reserveFactorScale);
-
-    return {
-        maxLiquidationAmount: liquidationAmount,
-        maxCollateralRewardAmount: collateralAmount
-    };
-}
-
-/**
- * This function shows how to calculate min collateral amount value.
- * when liquidator has not enough of loan asset to cover the full loan.
- * @param transferredAmount amount of loan asset liquidator want to transfer.
- * @param maxLiquidationAmount max liquidation amount value calculated.
- * @param maxCollateralReward max collateral reward amount calculated.
- * @returns minCollateralAmount value for safe liquidation.
- */
-export function calculateMinCollateralByTransferredAmount(
-    transferredAmount: bigint, maxLiquidationAmount: bigint, maxCollateralReward: bigint
-) {
-    if (maxLiquidationAmount === 0n) {
-        return 0n;
-    }
-    return maxCollateralReward * transferredAmount / maxLiquidationAmount;
 }
 
 /**
